LayoutTests/imported/w3c/web-platform-tests/selection/caret-position-should-be-correct-while-moveup-movedown-expected.txt

text1
text2
text1
text2
text1
text2

FAIL Caret position should be correct in moving up horizontal div when selection was left to right with line granularity assert_equals: expected 5 but got 0
PASS Caret position should be correct in moving up horizontal div when selection was right to left with line granularity
FAIL Caret position should be correct in moving down horizontal div when selection was left to right with line granularity assert_equals: expected 5 but got 0
PASS Caret position should be correct in moving down horizontal div when selection was right to left with line granularity
FAIL Caret position should be correct in moving up horizontal div when selection was left to right with paragraph granularity assert_equals: expected 5 but got 0
PASS Caret position should be correct in moving up horizontal div when selection was right to left with paragraph granularity
FAIL Caret position should be correct in moving down horizontal div when selection was left to right with paragraph granularity assert_equals: expected 5 but got 0
PASS Caret position should be correct in moving down horizontal div when selection was right to left with paragraph granularity
FAIL Caret position should be correct in move right with line granularity for vertical-lr div when selection was top to bottom assert_equals: expected 12 but got 7
PASS Caret position should be correct in move right with line granularity for vertical-lr div when selection was bottom to top
FAIL Caret position should be correct in move left with line granularity for vertical-lr div when selection was top to bottom assert_equals: expected 12 but got 10
FAIL Caret position should be correct in move left with line granularity for vertical-lr div when selection was bottom to top assert_equals: expected 7 but got 10
FAIL Caret position should be correct in move right with paragraph granularity for vertical-lr div when selection was top to bottom assert_equals: expected 12 but got 7
PASS Caret position should be correct in move right with paragraph granularity for vertical-lr div when selection was bottom to top
FAIL Caret position should be correct in move left with paragraph granularity for vertical-lr div when selection was top to bottom assert_equals: expected 12 but got 10
FAIL Caret position should be correct in move left with paragraph granularity for vertical-lr div when selection was bottom to top assert_equals: expected 7 but got 10
FAIL Caret position should be correct in move left with line granularity for vertical-rl div when selection was top to bottom assert_equals: expected 12 but got 7
PASS Caret position should be correct in move left with line granularity for vertical-rl div when selection was bottom to top
FAIL Caret position should be correct in move right with line granularity for vertical-rl div when selection was top to bottom assert_equals: expected 12 but got 10
FAIL Caret position should be correct in move right with line granularity for vertical-rl div when selection was bottom to top assert_equals: expected 7 but got 10
FAIL Caret position should be correct in move left with paragraph granularity for vertical-rl div when selection was top to bottom assert_equals: expected 12 but got 7
PASS Caret position should be correct in move left with paragraph granularity for vertical-rl div when selection was bottom to top
FAIL Caret position should be correct in move right with paragraph granularity for vertical-rl div when selection was top to bottom assert_equals: expected 12 but got 10
FAIL Caret position should be correct in move right with paragraph granularity for vertical-rl div when selection was bottom to top assert_equals: expected 7 but got 10
